Split service handles into files.
