+++ overwatch/src/services/mod.rs
@@ -1,13 +1,15 @@
 pub mod handle;
 pub mod life_cycle;
 pub mod relay;
+mod runner;
 pub mod settings;
 pub mod state;
+pub mod state_handle;
 pub mod status;
 
 use async_trait::async_trait;
-use handle::ServiceStateHandle;
 use state::ServiceState;
+use state_handle::ServiceStateHandle;
 
 /// The core data a service needs to handle.
 /// Holds the necessary information of a service.
